refactor: add CardInputWrapper
